+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 Unreleased
 
 * Refactored installation and packaging operations
+* New component build steps support
 * New `Directory` package source
 * New `Filesystem` package repository